FIX counter showing wrong numbers

+++ resources/js/src/app/components/item/AddToWishList.js
@@ -12,7 +12,8 @@ Vue.component("add-to-wish-list", {
     data()
     {
         return {
-            wishListCount: 0
+            wishListCount: 0,
+            isLoading: false
         };
     },
 
@@ -44,24 +45,42 @@ Vue.component("add-to-wish-list", {
 
         addToWishList()
         {
-            ApiService.post("/rest/io/itemWishList", {variationId: this.variationId})
-                .done(function()
-                {
-                    this.isActive = true;
-                    this.changeTooltipText();
-                    this.updateWatchListCount(parseInt(this.wishListCount.count) + 1);
-                }.bind(this));
+            if (!this.isLoading)
+            {
+                this.isLoading = true;
+                ApiService.post("/rest/io/itemWishList", {variationId: this.variationId})
+                    .done(() =>
+                    {
+                        this.isActive = true;
+                        this.isLoading = false;
+                        this.changeTooltipText();
+                        this.updateWatchListCount(parseInt(this.wishListCount.count) + 1);
+                    })
+                    .fail(() =>
+                    {
+                        this.isLoading = false;
+                    });
+            }
         },
 
         removeFromWishList()
         {
-            ApiService.delete("/rest/io/itemWishList/" + this.variationId)
-                .done(function()
-                {
-                    this.isActive = false;
-                    this.changeTooltipText();
-                    this.updateWatchListCount(parseInt(this.wishListCount.count) - 1);
-                }.bind(this));
+            if (!this.isLoading)
+            {
+                this.isLoading = true;
+                ApiService.delete("/rest/io/itemWishList/" + this.variationId)
+                    .done(() =>
+                    {
+                        this.isActive = false;
+                        this.isLoading = false;
+                        this.changeTooltipText();
+                        this.updateWatchListCount(parseInt(this.wishListCount.count) - 1);
+                    })
+                    .fail(() =>
+                    {
+                        this.isLoading = false;
+                    });
+            }
         },
 
         changeTooltipText()
